gosh: make tab complete optional at build
