+++ cmds/core/gosh/gosh.go
@@ -24,19 +24,13 @@ import (
 
 	"golang.org/x/term"
 
-	"github.com/knz/bubbline"
-
 	"mvdan.cc/sh/v3/interp"
 	"mvdan.cc/sh/v3/syntax"
 )
 
-const HISTFILE = "/tmp/bubble-sh.history" //TODO: make configurable
-
 func main() {
-	completion := flag.Bool("comp", false, "Enable tabcompletion and a more feature rich editline implementation")
-
 	flag.Parse()
-	err := run(os.Stdin, os.Stdout, os.Stderr, *completion, flag.Args()...)
+	err := run(os.Stdin, os.Stdout, os.Stderr, flag.Args()...)
 
 	if status, ok := interp.IsExitStatus(err); ok {
 		os.Exit(int(status))
@@ -47,7 +41,9 @@ func main() {
 	}
 }
 
-func run(stdin io.Reader, stdout, stderr io.Writer, completion bool, args ...string) error {
+var errNotImplemented = errors.New("fancy interactive interpreter not implemented")
+
+func run(stdin io.Reader, stdout, stderr io.Writer, args ...string) error {
 	runner, err := interp.New(interp.StdIO(stdin, stdout, stderr))
 	if err != nil {
 		return err
@@ -64,8 +60,8 @@ func run(stdin io.Reader, stdout, stderr io.Writer, completion bool, args ...str
 	}
 
 	if r, ok := stdin.(*os.File); ok && term.IsTerminal(int(r.Fd())) {
-		if completion {
-			return runInteractive(runner, parser, stdout, stderr)
+		if err := runInteractive(runner, parser, stdout, stderr); !errors.Is(err, errNotImplemented) {
+			return err
 		}
 		return runInteractiveSimple(runner, parser, stdin, stdout)
 	}
@@ -158,98 +154,4 @@ func runInteractiveSimple(runner *interp.Runner, parser *syntax.Parser, stdin io
 			return nil
 		}
 	}
-}
-
-func runInteractive(runner *interp.Runner, parser *syntax.Parser, stdout, stderr io.Writer) error {
-	input := bubbline.New()
-	// Set default window size to 80x24 in case ioctl isn't able to detect the actual window size
-	input.Model.SetSize(80, 24)
-
-	if err := input.LoadHistory(HISTFILE); err != nil {
-		return err
-	}
-
-	input.SetAutoSaveHistory(HISTFILE, true)
-
-	input.AutoComplete = autocomplete
-
-	var runErr error
-
-	// The following code is used to intercept SIGINT signals.
-	// Calling signal.Ignore wouldn't work as child prcesses inherit this trait.
-	// We only want to catch SIGINTs that are propagated from a child,
-	// the child itself should get the signal as per usual.
-	ch := make(chan os.Signal, 1)
-	signal.Notify(ch, os.Interrupt)
-	go func(ch chan os.Signal) {
-		for {
-			<-ch
-		}
-	}(ch)
-
-	for {
-		if runErr != nil {
-			fmt.Fprintf(stdout, "error: %s\n", runErr.Error())
-			runErr = nil
-		}
-
-		line, err := input.GetLine()
-
-		if err != nil {
-			if err == io.EOF {
-				break // maybe we should continue instead of break
-			}
-			if errors.Is(err, bubbline.ErrInterrupted) {
-				fmt.Fprintf(stdout, "^C\n")
-			} else {
-				fmt.Fprintf(stderr, "error: %s\n", err.Error())
-			}
-			err = nil
-			continue
-		}
-
-		switch line {
-		case "exit":
-			goto exit
-		case "disablecomp":
-			input.AutoComplete = nil
-			continue
-		case "enablecomp":
-			input.AutoComplete = autocomplete
-			continue
-		default:
-		}
-
-		// check if we want to execute a shell script
-		fields := strings.Fields(line)
-		if len(fields) > 0 && strings.HasSuffix(fields[0], "sh") {
-			if err := runScript(runner, parser, fields[0]); err != nil {
-				fmt.Fprintf(stderr, "error: %s\n", err.Error())
-			}
-
-			continue
-		}
-
-		if line != "" {
-			if err := input.AddHistory(line); err != nil {
-				fmt.Fprintf(stdout, "unable to add %s to history: %v\n", line, err)
-			}
-		}
-
-		if err := parser.Stmts(strings.NewReader(line), func(stmt *syntax.Stmt) bool {
-			if parser.Incomplete() {
-				fmt.Fprintf(stdout, "-> ")
-
-				return true
-			}
-
-			runErr = runner.Run(context.Background(), stmt)
-
-			return !runner.Exited()
-		}); err != nil {
-			fmt.Fprintf(stderr, "error: %s\n", err.Error())
-		}
-	}
-exit:
-	return nil
 }
